MONGOCRYPT-845 rename mlib directory to avoid header conflicts
