+++ CMakeLists.txt
@@ -233,22 +233,22 @@ configure_file (
    "${PROJECT_BINARY_DIR}/src/mongocrypt-config.h"
 )
 
-# Define the mlib target, which is private and header-only. It is not exported
+# Define the mc-mlib target, which is private and header-only. It is not exported
 # nor are its headers installed.
-add_library (_mongo-mlib INTERFACE)
-add_library (mongo::mlib ALIAS _mongo-mlib)
-list (APPEND MLIB_DEFINITIONS MLIB_USER)
+add_library (_mongo-mc-mlib INTERFACE)
+add_library (mongo::mc-mlib ALIAS _mongo-mc-mlib)
+list (APPEND MCMLIB_DEFINITIONS MLIB_USER)
 CHECK_INCLUDE_FILE (strings.h HAVE_STRINGS_H)
 if (HAVE_STRINGS_H)
-   list (APPEND MLIB_DEFINITIONS MLIB_HAVE_STRINGS_H)
+   list (APPEND MCMLIB_DEFINITIONS MLIB_HAVE_STRINGS_H)
 endif ()
 set_property(
-   TARGET _mongo-mlib
+   TARGET _mongo-mc-mlib
    APPEND PROPERTY INTERFACE_COMPILE_DEFINITIONS
-   ${MLIB_DEFINITIONS}
+   ${MCMLIB_DEFINITIONS}
    )
 set_property(
-   TARGET _mongo-mlib
+   TARGET _mongo-mc-mlib
    APPEND PROPERTY INTERFACE_INCLUDE_DIRECTORIES
    "$<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/src>"
    )
@@ -271,7 +271,7 @@ target_link_libraries (
    PRIVATE
       _mongocrypt::libbson_for_shared
       kms_message_static
-      $<BUILD_INTERFACE:mongo::mlib>
+      $<BUILD_INTERFACE:mongo::mc-mlib>
    PUBLIC
       mongocrypt::platform
       ${maybe_dfp_library}
@@ -327,7 +327,7 @@ target_link_libraries (
    PRIVATE
       _mongocrypt::libbson_for_static
       kms_message_static
-      $<BUILD_INTERFACE:mongo::mlib>
+      $<BUILD_INTERFACE:mongo::mc-mlib>
    PUBLIC
       mongocrypt::platform
       ${maybe_dfp_library}
@@ -507,7 +507,7 @@ set (TEST_MONGOCRYPT_SOURCES
 add_executable (test-mongocrypt ${TEST_MONGOCRYPT_SOURCES})
 # Use the static version since it allows the test binary to use private symbols
 target_include_directories (test-mongocrypt PRIVATE ./src "${CMAKE_CURRENT_SOURCE_DIR}/kms-message/src")
-target_link_libraries (test-mongocrypt PRIVATE _mongocrypt::libbson_for_static mongocrypt_static mongo::mlib)
+target_link_libraries (test-mongocrypt PRIVATE _mongocrypt::libbson_for_static mongocrypt_static mongo::mc-mlib)
 target_include_directories (test-mongocrypt PRIVATE "${CMAKE_CURRENT_LIST_DIR}/test")
 target_compile_definitions (test-mongocrypt PRIVATE
    # Set a definition so that testcases can know where test-mongocrypt.exe was written to
@@ -525,9 +525,9 @@ add_test (
    )
 
 foreach (test IN ITEMS path str)
-   add_executable (mlib.${test}.test src/mlib/${test}.test.c)
+   add_executable (mlib.${test}.test src/mc-mlib/${test}.test.c)
    add_test (mlib.${test} mlib.${test}.test)
-   target_link_libraries (mlib.${test}.test PRIVATE mongo::mlib)
+   target_link_libraries (mlib.${test}.test PRIVATE mongo::mc-mlib)
 endforeach ()
 
 if ("cxx_relaxed_constexpr" IN_LIST CMAKE_CXX_COMPILE_FEATURES)
@@ -545,7 +545,7 @@ if ("cxx_relaxed_constexpr" IN_LIST CMAKE_CXX_COMPILE_FEATURES)
       target_compile_features ("${exe_name}" PRIVATE cxx_relaxed_constexpr)
       target_link_libraries ("${exe_name}" PRIVATE
          mongocrypt
-         mongo::mlib
+         mongo::mc-mlib
          ${maybe_dfp_library}
          _mongocrypt::libbson_for_static
          )
@@ -555,7 +555,7 @@ endif ()
 
 if ("cxx_std_20" IN_LIST CMAKE_CXX_COMPILE_FEATURES)
    add_executable (csfle-markup src/csfle-markup.cpp)
-   target_link_libraries (csfle-markup PRIVATE mongocrypt_static _mongocrypt::libbson_for_static mongo::mlib)
+   target_link_libraries (csfle-markup PRIVATE mongocrypt_static _mongocrypt::libbson_for_static mongo::mc-mlib)
    target_compile_features (csfle-markup PRIVATE cxx_std_20)
 endif ()
 
@@ -606,8 +606,8 @@ install (
    INCLUDES DESTINATION ${CMAKE_INSTALL_INCLUDEDIR}
 )
 
-# This export set is not installed, and is only to allow export() of the mlib-using targets
-install (TARGETS _mongo-mlib EXPORT _exports_for_export)
+# This export set is not installed, and is only to allow export() of the mc-mlib-using targets
+install (TARGETS _mongo-mc-mlib EXPORT _exports_for_export)
 export (EXPORT _exports_for_export)
 
 install (
