+++ src/mc-dec128.test.cpp
@@ -6,7 +6,7 @@
 
 #include <stdlib.h>
 
-#include <mlib/check.hpp>
+#include <mc-mlib/check.hpp>
 #define CHECK MLIB_CHECK
 
 inline std::ostream &operator<<(std::ostream &out, mc_dec128 d) noexcept {
