+++ src/mc-mlib/str.h
