+++ src/mc-mlib/str.test.c
