+++ src/mc-range-encoding-private.h
@@ -21,7 +21,7 @@
 #include "mc-optional-private.h"
 #include "mongocrypt-status-private.h"
 
-#include <mlib/int128.h>
+#include <mc-mlib/int128.h>
 
 #include <stdint.h>
